Fixed imports, Improved renaming functions and constants
