+++ app/src/main/java/com/developersbreach/composeactors/ui/actorDetails/ActorDetailsViewModel.kt
@@ -5,10 +5,10 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.developersbreach.composeactors.model.ActorDetail
-import com.developersbreach.composeactors.model.Movie
-import com.developersbreach.composeactors.model.MovieDetail
-import com.developersbreach.composeactors.repository.network.NetworkRepository
+import com.developersbreach.composeactors.data.model.ActorDetail
+import com.developersbreach.composeactors.data.model.Movie
+import com.developersbreach.composeactors.data.model.MovieDetail
+import com.developersbreach.composeactors.repository.NetworkRepository
 import kotlinx.coroutines.launch
 import timber.log.Timber
 import java.io.IOException
