+++ app/src/main/java/com/developersbreach/composeactors/ui/modalSheet/SheetContentActorDetails.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.developersbreach.composeactors.R
-import com.developersbreach.composeactors.model.ActorDetail
+import com.developersbreach.composeactors.data.model.ActorDetail
 import com.developersbreach.composeactors.ui.actorDetails.ActorInfoHeader
 import com.developersbreach.composeactors.ui.components.LoadNetworkImage
