+++ app/src/main/java/com/developersbreach/composeactors/ui/modalSheet/SheetContentMovieDetails.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.developersbreach.composeactors.R
-import com.developersbreach.composeactors.model.MovieDetail
+import com.developersbreach.composeactors.data.model.MovieDetail
 import com.developersbreach.composeactors.ui.components.LoadNetworkImage
 import com.developersbreach.composeactors.ui.movieDetail.MovieGenre
