+++ app/src/main/java/com/developersbreach/composeactors/ui/movieDetail/MovieDetailViewModel.kt
@@ -6,12 +6,12 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.developersbreach.composeactors.model.ActorDetail
-import com.developersbreach.composeactors.model.Cast
-import com.developersbreach.composeactors.model.Movie
-import com.developersbreach.composeactors.model.MovieDetail
-import com.developersbreach.composeactors.repository.database.DatabaseRepository
-import com.developersbreach.composeactors.repository.network.NetworkRepository
+import com.developersbreach.composeactors.data.model.ActorDetail
+import com.developersbreach.composeactors.data.model.Cast
+import com.developersbreach.composeactors.data.model.Movie
+import com.developersbreach.composeactors.data.model.MovieDetail
+import com.developersbreach.composeactors.repository.DatabaseRepository
+import com.developersbreach.composeactors.repository.NetworkRepository
 import kotlinx.coroutines.launch
 import timber.log.Timber
 import java.io.IOException
